feat: Split work order form into smaller components (DSD-96)
