+++ src/components/schedule/no-times-avail.tsx
@@ -1,0 +1,12 @@
+export default function NoTimesAvailable() {
+  return (
+    <div className="pb-4">
+      <span className="block text-center text-black">
+        Unfortunately, no times are available on this day.
+      </span>
+      <span className="block text-center">
+        Please go back and select another day.
+      </span>
+    </div>
+  );
+}
