+++ src/components/schedule/today-selected.tsx
@@ -1,0 +1,13 @@
+export default function TodaySelected() {
+  return (
+    <div className="pb-4 text-center">
+      <span className="font-medium">You selected today.</span>
+      <span className="block">
+        If you need a same-day appointment,{" "}
+        <span className="block">
+          please call our office to check availability.
+        </span>
+      </span>
+    </div>
+  );
+}
